Merge with master

+++ Setup/ericomshield-uninstall.sh
@@ -40,8 +40,9 @@ docker swarm leave -f
 
 docker system prune -f -a
 
-rm -f $ES_PATH/*
-rm -f $ES_PATH/.*
-
-
+if [ "$1" == "-a" ]; then
+  rm -rf $ES_PATH
+ else
+  rm -f $ES_PATH/*
+fi
 echo "Done!"
